refactor: split accelerator backend into accelerators package
