+++ modelexpress_client/python/modelexpress/accelerators/base.py
@@ -1,11 +1,10 @@
 # SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Accelerator backend abstraction for device-specific operations."""
+"""Accelerator backend boundary for device-specific operations."""
 
 from __future__ import annotations
 
-from dataclasses import dataclass
 from typing import Protocol
 
 import torch
@@ -71,64 +70,3 @@ class AcceleratorBackend(Protocol):
     def supports_gds(self) -> bool:
         """Return whether GPUDirect Storage loading is supported."""
         ...
-
-
-@dataclass(frozen=True)
-class CudaAcceleratorBackend:
-    """CUDA implementation of the accelerator backend boundary."""
-
-    @property
-    def name(self) -> str:
-        return "cuda"
-
-    @property
-    def torch_device_type(self) -> str:
-        return "cuda"
-
-    @property
-    def nixl_mem_type(self) -> str:
-        return NIXL_ACCELERATOR_MEM_TYPE
-
-    def set_device(self, device_id: int) -> None:
-        torch.cuda.set_device(device_id)
-
-    def current_device(self) -> int:
-        return int(torch.cuda.current_device())
-
-    def synchronize(self, device_id: int | None = None) -> None:
-        if device_id is None:
-            torch.cuda.synchronize()
-        else:
-            torch.cuda.synchronize(device_id)
-
-    def empty_cache(self) -> None:
-        torch.cuda.empty_cache()
-
-    def torch_device(self, device_id: int) -> torch.device:
-        return torch.device(self.torch_device_type, device_id)
-
-    def is_accel_tensor(self, tensor: torch.Tensor) -> bool:
-        return bool(tensor.is_cuda)
-
-    def supports_rdma_p2p(self) -> bool:
-        return True
-
-    def supports_pool_reg(self) -> bool:
-        return True
-
-    def supports_vmm(self) -> bool:
-        return True
-
-    def supports_gds(self) -> bool:
-        return True
-
-
-def accelerator_backend_for(device: torch.device | str) -> AcceleratorBackend:
-    """Return the backend implementation for ``device``.
-
-    Only CUDA devices are currently supported by this factory.
-    """
-    torch_device = torch.device(device)
-    if torch_device.type == "cuda":
-        return CudaAcceleratorBackend()
-    raise ValueError(f"Unsupported accelerator backend for torch device {torch_device!s}")
