+++ modelexpress_client/python/modelexpress/engines/sglang/adapter.py
@@ -16,7 +16,7 @@ import torch
 
 from ... import p2p_pb2
 from ...adapter import EngineAdapter
-from ...accelerator_backend import accelerator_backend_for
+from ...accelerators import accelerator_backend_for
 from ...load_strategy.context import LoadContext, LoadResult
 from ...metadata.client_factory import create_metadata_client
