+++ modelexpress_client/python/tests/test_nixl_backend.py
@@ -5,7 +5,7 @@
 
 import pytest
 
-from modelexpress.accelerator_backend import NIXL_ACCELERATOR_MEM_TYPE
+from modelexpress.accelerators import NIXL_ACCELERATOR_MEM_TYPE
 from modelexpress.nixl_transfer import (
     DEFAULT_NIXL_BACKEND,
     NixlTransferManager,
